Ready version that uses redis data store
We used redis to hold the data for users in the room, used latest
version of socketio. An admin user is now used to manage persisting
sessions for up to 24hrs after the meeting has ended as opposed to
deleting it immediately which was the previous logic. Added the redis
dependency to the requirement.txt file.